Updating for llvm 3.1 and Ocaml 4.0
Part of the update is trivial, just changing the version numbers in configure script.

Adapting to the type handling change in llv3 3.0 was a bit more involved, mostly because I needed to figure out what is happennig. But the change is rather simple 
and straigthforward.

The obsoleted command options for llvm opt are just commented out, because I'm
not sure what their equivalents are.